refactor: rename create-application-harness
